+++ apps/kyb-app/src/components/organisms/UIRenderer/elements/SubmitButton/SubmitButton.tsx
@@ -12,15 +12,16 @@ import { UIElementComponent } from '@/components/organisms/UIRenderer/types';
 import { UIPage } from '@/domains/collection-flow';
 import { useFlowTracking } from '@/hooks/useFlowTracking';
 import { Button } from '@ballerine/ui';
-import { useCallback } from 'react';
+import { useCallback, useMemo } from 'react';
 
 export const SubmitButton: UIElementComponent<{ text: string }> = ({ definition }) => {
   const { helpers } = useDynamicUIContext();
   const { onClickHandler } = useUIElementHandlers(definition);
   const { state } = useDynamicUIContext();
   const { state: uiElementState } = useUIElementState(definition);
-  const { currentPage } = usePageResolverContext();
+  const { currentPage, pages } = usePageResolverContext();
   const { errors } = usePageContext();
+  const isValid = useMemo(() => !Object.values(errors).length, [errors]);
 
   const setPageElementsTouched = useCallback(
     (page: UIPage, state: UIState) => {
@@ -64,8 +65,13 @@ export const SubmitButton: UIElementComponent<{ text: string }> = ({ definition 
       state,
     );
     onClickHandler();
-    trackFinish();
-  }, [currentPage, state, setPageElementsTouched, onClickHandler, trackFinish]);
+
+    const isFinishPage = currentPage?.name === pages.at(-1)?.name;
+
+    if (isFinishPage && isValid) {
+      trackFinish();
+    }
+  }, [currentPage, pages, state, isValid, setPageElementsTouched, onClickHandler, trackFinish]);
 
   return (
     <Button