+++ services/workflows-service/scripts/workflows/generate-base-case-level-states.ts
@@ -3,31 +3,58 @@ import { CommonWorkflowEvent, CommonWorkflowStates, StateTag } from '@ballerine/
 export const generateBaseCaseLevelStates = (
   defaultState: string = CommonWorkflowStates.MANUAL_REVIEW,
   defaultResubmitEvent: string = CommonWorkflowEvent.RETURN_TO_REVIEW,
-) => ({
-  [defaultState]: {
-    tags: [StateTag.MANUAL_REVIEW],
-    on: {
-      [CommonWorkflowEvent.REJECT]: { target: CommonWorkflowStates.REJECTED },
-      [CommonWorkflowEvent.APPROVE]: { target: CommonWorkflowStates.APPROVED },
-      [CommonWorkflowEvent.REVISION]: { target: CommonWorkflowStates.REVISION },
+) =>
+  ({
+    [defaultState]: {
+      tags: [StateTag.MANUAL_REVIEW],
+      on: {
+        [CommonWorkflowEvent.REJECT]: { target: CommonWorkflowStates.REJECTED },
+        [CommonWorkflowEvent.APPROVE]: { target: CommonWorkflowStates.APPROVED },
+        [CommonWorkflowEvent.REVISION]: { target: CommonWorkflowStates.REVISION },
+      },
     },
-  },
-  rejected: {
-    tags: [StateTag.REJECTED],
-    type: 'final',
-  },
-  approved: {
-    tags: [StateTag.APPROVED],
-    type: 'final',
-  },
-  resolved: {
-    tags: [StateTag.RESOLVED],
-    type: 'final',
-  },
-  revision: {
-    tags: [StateTag.REVISION],
-    on: {
-      [defaultResubmitEvent]: defaultState,
+    [CommonWorkflowStates.REJECTED]: {
+      tags: [StateTag.REJECTED],
+      type: 'final',
     },
-  },
-});
+    [CommonWorkflowStates.APPROVED]: {
+      tags: [StateTag.APPROVED],
+      type: 'final',
+    },
+    [CommonWorkflowStates.RESOLVED]: {
+      tags: [StateTag.RESOLVED],
+      type: 'final',
+    },
+    [CommonWorkflowStates.REVISION]: {
+      tags: [StateTag.REVISION],
+      on: {
+        [defaultResubmitEvent]: defaultState,
+      },
+    },
+  } as const);
+
+export const generateBaseCaseLevelStatesAutoTransitionOnRevision = (
+  defaultState: string = CommonWorkflowStates.MANUAL_REVIEW,
+  defaultResubmitEvent: string = CommonWorkflowEvent.RETURN_TO_REVIEW,
+) => {
+  const definition = generateBaseCaseLevelStates(defaultState, defaultResubmitEvent);
+
+  return {
+    ...definition,
+    [CommonWorkflowStates.REVISION]: {
+      ...definition[CommonWorkflowStates.REVISION],
+      on: {
+        ...definition[CommonWorkflowStates.REVISION].on,
+      },
+      always: {
+        target: defaultState,
+        cond: {
+          type: 'jmespath',
+          options: {
+            rule: 'length(documents[?decision.status]) < length(documents)',
+          },
+        },
+      },
+    },
+  };
+};