+++ services/workflows-service/src/end-user/end-user.controller.external.ts
@@ -8,7 +8,7 @@ import type { Request } from 'express';
 
 import { ApiNestedQuery } from '@/common/decorators/api-nested-query.decorator';
 import * as errors from '../errors';
-import * as nestAccessControl from 'nest-access-control';
+// import * as nestAccessControl from 'nest-access-control';
 import { EndUserCreateDto } from './dtos/end-user-create';
 import { EndUserFindManyArgs } from './dtos/end-user-find-many-args';
 import { EndUserWhereUniqueInput } from './dtos/end-user-where-unique-input';
@@ -29,9 +29,7 @@ import { CurrentProject } from '@/common/decorators/current-project.decorator';
 export class EndUserControllerExternal {
   constructor(
     protected readonly service: EndUserService,
-    protected readonly workflowService: WorkflowService,
-    @nestAccessControl.InjectRolesBuilder()
-    protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
+    protected readonly workflowService: WorkflowService, // @nestAccessControl.InjectRolesBuilder() // protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
   ) {}
 
   @common.Post()
@@ -59,6 +57,7 @@ export class EndUserControllerExternal {
         avatarUrl: true,
       },
     });
+
     return endUser;
   }
 
@@ -92,6 +91,7 @@ export class EndUserControllerExternal {
     @ProjectIds() projectIds: TProjectIds,
   ): Promise<EndUserModel[]> {
     const args = plainToClass(EndUserFindManyArgs, request.query);
+
     return this.service.list(args, projectIds);
   }