+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/eslint-config-react
 
+## 1.0.8
+
+### Patch Changes
+
+- version bump
+- added rules to enforce spacing
+- Updated dependencies
+  - @ballerine/eslint-config@1.0.7
+
+## 1.0.7
+
+### Patch Changes
+
+- Ver bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.0.6
+
 ## 1.0.6
 
 ### Patch Changes