+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.5.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.47
+  - @ballerine/common@0.7.47
+
+## 0.5.46
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.46
+  - @ballerine/workflow-core@0.5.46
+
+## 0.5.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-core@0.5.45
+
+## 0.5.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.44
+  - @ballerine/common@0.7.44
+
+## 0.5.43
+
+### Patch Changes
+
+- Ver bump
+- Updated dependencies
+  - @ballerine/common@0.7.43
+  - @ballerine/workflow-core@0.5.43
+
+## 0.5.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.42
+  - @ballerine/workflow-core@0.5.42
+
 ## 0.5.41
 
 ### Patch Changes